requirements.txt
discord.py==2.4.0
yfinance>=1.2.0
aiosqlite==0.20.0
python-dotenv==1.0.1
pytz==2024.1
